Add support for powershell in stack shellinit
